+++ tests/estimators/mutual_information/test_renyi.py
@@ -131,3 +131,33 @@ def test_renyi_cmi_autoregressive(rng_int, k, alpha, expected):
         data_x, data_y, cond=cond, k=k, alpha=alpha, base=2, noise_level=0
     )
     assert est.result() == pytest.approx(expected)
+
+
+@pytest.mark.parametrize(
+    "rng_int,method,p_mi,p_cmi",
+    [
+        (1, "permutation_test", 0.4, 0.08),
+        (1, "bootstrap", 0.0, 0.04),
+        (2, "permutation_test", 0.1, 0.32),
+        (2, "bootstrap", 0.0, 0.22),
+        (3, "permutation_test", 0.56, 0.2),
+        (3, "bootstrap", 0.0, 0.02),
+        (4, "permutation_test", 0.1, 0.64),
+        (4, "bootstrap", 0.0, 0.32),
+    ],
+)
+def test_renyi_mi_statistical_test(rng_int, method, p_mi, p_cmi):
+    """Test the Renyi MI for p-values. Fix rng."""
+    data_x, data_y, cond = generate_autoregressive_series_condition(
+        rng_int, alpha=(0.5, 0.1), beta=0.6, gamma=(0.4, 0.2)
+    )
+    est_mi = RenyiMIEstimator(
+        data_x, data_y, k=4, alpha=1.0, noise_level=0, base=2, seed=8
+    )
+    est_cmi = RenyiCMIEstimator(
+        data_x, data_y, cond=cond, k=4, alpha=1.0, noise_level=0, base=2, seed=8
+    )
+    test = est_mi.statistical_test(method=method, n_tests=50)
+    assert test.p_value == pytest.approx(p_mi)
+    test = est_cmi.statistical_test(method=method, n_tests=50)
+    assert test.p_value == pytest.approx(p_cmi)