+++ tests/estimators/transfer_entropy/test_miller_madow.py
@@ -225,3 +225,53 @@ def test_miller_madow_te_uncoupled(default_rng):
     res_xy = est_xy.result()
     assert isinstance(res_xy, float)
     assert res_xy == pytest.approx(0.0, abs=0.1)
+
+
+@pytest.mark.parametrize(
+    "rng_int,method,p_te,p_cte",
+    [
+        (1, "permutation_test", 0.0, 0.0),
+        (1, "bootstrap", 0.0, 0.0),
+        (2, "permutation_test", 0.0, 0.0),
+        (2, "bootstrap", 0.0, 0.0),
+        (3, "permutation_test", 0.0, 0.0),
+        (4, "permutation_test", 0.0, 0.0),
+    ],
+)
+def test_miller_madow_te_statistical_test(rng_int, method, p_te, p_cte):
+    """Test the Miller-Madow TE for p-values. Fix rng."""
+    data_source, data_dest, data_cond = discrete_random_variables_condition(rng_int)
+    est_te_xy = MillerMadowTEEstimator(data_source, data_dest, base=2, seed=8)
+    est_cte_xy = MillerMadowCTEEstimator(
+        data_source, data_dest, cond=data_cond, base=2, seed=8
+    )
+    test = est_te_xy.statistical_test(method=method, n_tests=50)
+    assert test.p_value == pytest.approx(p_te)
+    test = est_cte_xy.statistical_test(method=method, n_tests=50)
+    assert test.p_value == pytest.approx(p_cte)
+
+
+@pytest.mark.parametrize(
+    "rng_int,method,eff_te,eff_cte",
+    [
+        (1, "permutation_test", 1.0012712760043667, 0.0),
+        (1, "bootstrap", 1.0074631039986939, 0.0),
+        (2, "permutation_test", 0.9992019340962991, 0.0),
+        (2, "bootstrap", 1.0116229970679638, 0.0),
+        (3, "permutation_test", 1.0024608869442286, 0.0),
+        (4, "permutation_test", 0.9962627039837995, 0.0),
+    ],
+)
+def test_miller_madow_te_effective_val(rng_int, method, eff_te, eff_cte):
+    """Test the Miller-Madow transfer entropy for effective values. Fix rng."""
+    data_source, data_dest, data_cond = discrete_random_variables_condition(rng_int)
+    est_te_xy = MillerMadowTEEstimator(data_source, data_dest, base=2, seed=8)
+    est_cte_xy = MillerMadowCTEEstimator(
+        data_source,
+        data_dest,
+        cond=data_source,
+        base=2,
+        seed=8,
+    )
+    assert est_te_xy.effective_val(method=method) == pytest.approx(eff_te)
+    assert est_cte_xy.effective_val(method=method) == pytest.approx(eff_cte)